requirements.txt
python-dotenv==0.10.3
requests==2.22.0
terminaltables==3.1.0